+++ src/components/carts/CartsDelete.jsx
@@ -1,44 +1,15 @@
 import PropTypes from "prop-types";
-import useAxiosInstance from "@hooks/useAxiosInstance";
 
 export default function CartsDelete({
   handleAllCheckboxChange,
   allChecked,
-  checkedItems,
-  setCarts,
-  setLoading,
-  setError,
+  DeleteSelectedCarts,
 }) {
-  const axios = useAxiosInstance();
-
-  // 선택된 상품 장바구니에서 제거
-  const DeleteSelectedCarts = async () => {
-    if (checkedItems.length === 0) {
-      console.log("삭제할 항목이 없습니다.");
-      return;
-    }
+  const handleDelete = () => {
     const deleteCartsConfirm = window.confirm(
       "선택된 상품을 장바구니에서 제거 하시겠습니까?",
     );
     if (!deleteCartsConfirm) return;
-    setLoading(true); // 삭제 요청 시작 시 로딩 상태 설정
-    try {
-      await Promise.all(checkedItems.map(id => axios.delete(`/carts/${id}`))); // 여러 개 DELETE 요청
-      // 로컬 상태에서 해당 아이템 제거
-      setCarts(prevCarts => ({
-        ...prevCarts,
-        item: prevCarts.item.filter(cart => !checkedItems.includes(cart._id)), // 삭제된 아이템 제외
-      }));
-    } catch (err) {
-      setError(err);
-      console.error("선택 삭제 중 오류 발생:", err);
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  const handleDelete = () => {
-    console.log("선택 삭제가 실행됩니다.", checkedItems);
     DeleteSelectedCarts(); // 선택 삭제 함수 호출
   };
 
@@ -67,10 +38,7 @@ export default function CartsDelete({
 }
 
 CartsDelete.propTypes = {
-  checkedItems: PropTypes.array.isRequired,
   handleAllCheckboxChange: PropTypes.func.isRequired,
   allChecked: PropTypes.bool.isRequired,
-  setCarts: PropTypes.func.isRequired,
-  setLoading: PropTypes.func.isRequired,
-  setError: PropTypes.func.isRequired,
+  DeleteSelectedCarts: PropTypes.func.isRequired,
 };
